Audit core solver: fix extreme-coordinate overflow panics, cost-chain overflow, NaN scoring
Core audit (docs/audit-2026-08-12-core.md, 11 findings: 3 fixed /
6 reported / 2 folded into recommendations). Fixed with regression
tests (tests/input_boundary.rs, 7 tests that failed before):

- P1: extreme seat_positions (1e300) pass the finiteness check, saturate
  to +/-i32::MAX via round().as i32, then (row - col) style subtraction
  overflowed - debug panic, release silent wraparound. Eight sites in
  objectives/cost/scoring/reports now compute i64/f64 differences;
  normal-input semantics unchanged (aligned with Python's arbitrary
  precision).
- P1: height_cm 1e300 saturates i64::MAX and the cost chain overflowed
  when multiplied by row deltas; cost.rs now uses a saturating chain.
- P2: non-finite student score/height_cm were not validated in-process;
  a NaN score polluted percentiles and serialized total_cost as JSON
  null. engine.rs validates finiteness (JSON transport already rejects
  such values; this guards in-process DTO construction).

Reported (registered): no scale cap on validate's O(V^2) distance
matrix (~900 MB at 10k seats), extreme-score +/-1e308 diff overflow
to inf (JSON null), max_distance NaN/inf unvalidated, duplicate
seat_id/coordinates unvalidated, i32 saturation vs oracle numeric
divergence (tie to the export extent cap), student_count==0 semantics
mismatch. 80 lib + 7 boundary + integration tests green, clippy
clean.

+++ crates/seattrellis-core/src/reports.rs
@@ -365,8 +365,10 @@ pub fn pair_report_json(
                     *pair.relation_counts.entry(relation.clone()).or_default() += 1;
                     *relation_totals.entry(relation.clone()).or_default() += 1;
                 }
-                let row_delta = (first_seat.row - second_seat.row).unsigned_abs();
-                let col_delta = (first_seat.col - second_seat.col).unsigned_abs();
+                // Deltas in i64: the i32 subtraction of saturated extreme
+                // coordinates would overflow (debug panic).
+                let row_delta = (i64::from(first_seat.row) - i64::from(second_seat.row)).unsigned_abs();
+                let col_delta = (i64::from(first_seat.col) - i64::from(second_seat.col)).unsigned_abs();
                 pair.records.push(json!({
                     "snapshot_index": snapshot_index,
                     "first_seat_id": first_seat_id,
